Examine Page: Sort records by chromosome & postion
This change also sorts chromosomes in the dropdown chromosome select
box.

Filtered records are sorted on every ExaminePage render.

This change results in a performance hit on every re-render. This is
only noticable when sorting more than >20k or so records. With 40k
records, the penalty is about 700ms if sorting all 40k of them (an
uncommon case). Fixing this is probably worth doing, but doing it right
will have far-reaching consequences better relegated to a different
branch.

fixes #61
fixes #61