CQL QueryBox in the Examine page.
I've ripped out the filter boxes for individual columns in favor of the
monolithic search box. Typing "ORDER BY sample:GQ" will update the
sort arrows in the appropriate column. The inverse (clicking the arrow)
doesn't update the query (yet!).

This also renames a few things to be more uniform.