Derive Truth VCF for stats calculation
This fixes #251, deriving truth VCFs for calculation of TFPN stats on
the Examine page. If a validation VCF exists for a given dataset, the
first one returned by a query for such a VCF is inferred and used.

This fixes #267, so truth VCFs can be submitted.

This adds a badge to the runs table on truth VCFs, indicating to the
user that the VCF displayed there is a validation VCF.