Organize runs page by project
The projects are sorted so that the project with the most recent run is
at the top, and the runs are sorted by most recent at the top within
each project.

This is the first step to reorganizing the runs page, and one that
requires no changes to the database.