Disable surprising link swaps.
Fixes #523.

Here's what I think was happening: initially, the "Load" link has a
react ID of something like ".foo.1" and the "jump" link is ".foo.3".
When you click the link, React does a DOM update and swaps the links (so
that they become "Jump" and "reload"). Because the order has changed,
the "Jump" link is now ".foo.1" and the "reload" link is now ".foo.3".
React seems to only trigger the true DOM event at this point, after
doing an update cycle. So the link that's clicked is ".foo.1", which is
now the "Jump" link. I found this sequencing quite surprising (and hard
to debug!).

There are two fixes in this commit, either of which would have worked on
its own:

1. Adding explicit React keys to each link.
2. Disabling the on-screen swap when you click a link.